requirements.txt
requests
folium
textblob
wordcloud
matplotlib
Pillow
